Refactored gpt.V2 to separate generic methods.
